Add model_driver_PF_cubic_splines__MD_620624592962_002
Note that the format of the parameter files read by this driver has been
changed so that the chemical species is listed as a string on the first
line.  Although the nuclear charge (i.e. atomic number) is listed just
after the cutoff, KIM API v2 has no functions that work directly in
terms of this, but rather only character strings.