Add IMD driver
I'm still observing cases where the forces are not consistent with the
energy.  For example, hardcode FCCspacing to 5.5 Angstroms in the
utility_forces_numer_deriv example simulator and then run it against
IMD_EAM_Schopf_AlMnPd__MO_878712978062_003 and you'll see significant
disagreement.  This also appears for most all of the other IMD
parametrized Models.  The only one that appears to always have
consistent forces is IMD_EAM_Schopf_CaCd__MO_145183423516_003.  I
personally wonder whether this is due to some highly oscillatory
"ringing" in the spline interpolations that has a frequency higher than
that captured in the energy sampling done by the util.