Add hints to EDIP driver
1. I realized that I had forgotten a check on whether particle i is
contributing in the outermost loop.  There are no checks on whether the
neighbors of i are contributing, as the energy terms that are added on
are those designed for a full list summation. The force contributions
are also handled in a full list style, so no adjustment should be
needed.

2. For now, I've set halfListHints to 0.  However, I think it's possible
to make a more efficient looping structure that *might* permit half
lists if only the energy is being computed rather than the forces.  This
is future work.